Code fixes

+++ app/src/main/java/code/name/monkey/retromusic/adapter/song/SongAdapter.kt
@@ -60,7 +60,7 @@ open class SongAdapter(
     }
 
     override fun getItemId(position: Int): Long {
-        return dataSet[position].id.toLong()
+        return dataSet[position].id
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
